FlowerShopBELC/requirements.txt
fastapi
uvicorn
pydantic
tensorflow
opencv-python
numpy
pyodbc  
python-dotenv  
python-multipart